Feat : createOrder 개발
- createOrder 기능 개발
- createOrder 테스트 완료

+++ src/test/java/team/mosk/api/server/domain/options/option/util/GivenOption.java
@@ -7,8 +7,8 @@ public class GivenOption {
 
     public static final String MODIFIED_OPTION_NAME = "NEW 테스트 옵션";
 
-    public static final Long OPTION_PRICE = 1000L;
-    public static final Long MODIFIED_OPTION_PRICE = 10000L;
+    public static final int OPTION_PRICE = 1000;
+    public static final int MODIFIED_OPTION_PRICE = 10000;
 
     public static Option toEntity() {
         return Option.builder()
